requirements.txt
aio-pika==9.4.1
aiormq==6.8.0
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
asynctest==0.13.0
certifi==2024.2.2
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mirakuru==2.5.2
multidict==6.0.5
orjson==3.10.3
packaging==24.0
pamqp==3.3.0
pika==1.3.2
pluggy==1.5.0
port-for==0.7.2
psutil==5.9.8
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-extra-types==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.18.0
pytest==8.2.1
pytest-asyncio==0.23.7
pytest-rabbitmq==3.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
typer==0.12.3
typing_extensions==4.12.0
ujson==5.10.0
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
yarl==1.9.4
